+++ examples/xor.rs
@@ -10,31 +10,26 @@ pub fn main() {
         .with_output_type(ActivationType::Sigmoid); //we expect it to be 0 or 1 so sigmoid is ok
 
     let mut inp_out = [
-        ([0., 0.], [0.1]),
-        ([0., 1.], [1.]),
-        ([1., 0.], [1.]),
-        ([1., 1.], [0.]),
+        (vec![0., 0.], vec![0.1]),
+        (vec![0., 1.], vec![1.]),
+        (vec![1., 0.], vec![1.]),
+        (vec![1., 1.], vec![0.]),
     ];
 
     let mut mse_sum = 0.;
-    let mut avg_mse = 0.;
 
-    for e in 1..1000 {
+    for _ in 1..1000 {
         fastrand::shuffle(&mut inp_out);
 
-        let ins = inp_out.iter().map(|x| &x.0[..]).collect::<Vec<&[f32]>>();
-        let outs = inp_out.iter().map(|x| &x.1[..]).collect::<Vec<&[f32]>>();
+        let ins = inp_out.iter().map(|x| &x.0).collect::<Vec<_>>();
+        let outs = inp_out.iter().map(|x| &x.1).collect::<Vec<_>>();
 
         nn.fit(&ins, &outs); //we do batch descent, passing in all observations each time
-        let err: f32 = ins
-            .iter()
-            .zip(outs)
-            .map(|(ins, outs)| nn.forward_error(&ins, &outs))
-            .sum();
+        let mse: f32 = nn.forward_errors(&ins, &outs);
 
-        mse_sum += err;
-        avg_mse = mse_sum / e as f32;
+        mse_sum += mse;
     }
+    let avg_mse = mse_sum / 1000.;
 
     println!("avg mse: {avg_mse}");
     println!("We want 0,1,1,0:");