+++ src/lib.rs
@@ -1,4 +1,5 @@
 pub mod activation;
+pub mod dataset;
 pub mod initialization;
 pub mod nn;
 pub mod regularization;