+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+### 0.3.0
+- added Dataset manager enabling:
+    - one hot encoding (this enables classification)
+    - normalisation
+    - quickly read from csv
+- added iris example
+- added some utility functions:
+    - `forward_errors` to forward an entire batch inputs to get the MSE
+    - `max_index_equal` used with one hot encoding to calc the accuracy
+
 ### 0.2.0
 - added regularization L1, L2, L1 and L2
 - added examples: